Add QuicPacketCreator and QuicFecGroup.
Review URL: https://chromiumcodereview.appspot.com/11238013

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@163724 0039d316-1c4b-4281-b951-d872f2087c98

+++ net/quic/quic_utils.h
@@ -25,7 +25,7 @@ class NET_EXPORT_PRIVATE QuicUtils {
  public:
   // The overhead the quic framing will add for a packet with num_frames
   // frames.
-  static int StreamFramePacketOverhead(int num_frames);
+  static size_t StreamFramePacketOverhead(int num_frames);
 
   // returns the 128 bit FNV1a hash of the data.  See
   // http://www.isthe.com/chongo/tech/comp/fnv/index.html#FNV-param
